inference pt 1: symbolic potential vars in fn args
this changes goal generation such that when we create synthesis
goals from function declarations, if potential inference is enabled,
we automatically replace the potentials for all function arguments
and their type parameters with a symbolic potential variable which
will be inferred.

this does not attempt to replace abstract potentials with symbolic
abstract potentials (hopefully will be fixed in the future), and
this doesn't actually implement inference.